Create mac_add_charger.py
The following program will change the MAC address of a network interface using the "ifconfig" command. This program is part of one the projects from the Udemy course called Python & Ethical Hacking From Scratch. Some of the improvements that I made are the addition of input validation and make some corrections to be compatible with python 3.